requirements.txt
SQLAlchemy==2.0.27
python-dotenv==1.0.1
pyautogui==0.9.54
opencv-python==4.11.0.86
easyocr==1.7.1
playwright==1.41.2
Flask==3.0.2
pytest==8.0.1
python-logging-loki==0.3.1 
cryptography==44.0.3
mouse==0.7.1
pysqlcipher3==1.2.0
sqlite3==2.6.0
psutil==6.0.0
pillow==11.1.0
pywin32==306